+++ kernel-rs/src/virtio_disk.rs
@@ -190,7 +190,7 @@ unsafe fn free_desc(mut i: i32) {
     if disk.free[i as usize] != 0 {
         panic(b"virtio_disk_intr 2\x00" as *const u8 as *const libc::c_char as *mut libc::c_char);
     }
-    (*disk.desc.offset(i as isize)).addr = 0 as i32 as usize;
+    (*disk.desc.offset(i as isize)).addr = 0usize;
     disk.free[i as usize] = 1 as i32 as libc::c_char;
     wakeup(
         &mut *disk.free.as_mut_ptr().offset(0 as i32 as isize) as *mut libc::c_char
@@ -258,41 +258,41 @@ pub unsafe fn virtio_disk_rw(mut b: *mut Buf, mut write: i32) {
 
     // buf0 is on a kernel stack, which is not direct mapped,
     // thus the call to kvmpa().
-    (*disk.desc.offset(idx[0 as i32 as usize] as isize)).addr =
+    (*disk.desc.offset(idx[0usize] as isize)).addr =
         kvmpa(&mut buf0 as *mut virtio_blk_outhdr as usize);
-    (*disk.desc.offset(idx[0 as i32 as usize] as isize)).len =
+    (*disk.desc.offset(idx[0usize] as isize)).len =
         ::core::mem::size_of::<virtio_blk_outhdr>() as u32;
-    (*disk.desc.offset(idx[0 as i32 as usize] as isize)).flags = VRING_DESC_F_NEXT as u16;
-    (*disk.desc.offset(idx[0 as i32 as usize] as isize)).next = idx[1 as i32 as usize] as u16;
-    (*disk.desc.offset(idx[1 as i32 as usize] as isize)).addr = (*b).data.as_mut_ptr() as usize;
-    (*disk.desc.offset(idx[1 as i32 as usize] as isize)).len = BSIZE as u32;
+    (*disk.desc.offset(idx[0usize] as isize)).flags = VRING_DESC_F_NEXT as u16;
+    (*disk.desc.offset(idx[0usize] as isize)).next = idx[1usize] as u16;
+    (*disk.desc.offset(idx[1usize] as isize)).addr = (*b).data.as_mut_ptr() as usize;
+    (*disk.desc.offset(idx[1usize] as isize)).len = BSIZE as u32;
     if write != 0 {
         // device writes b->data
-        (*disk.desc.offset(idx[1 as i32 as usize] as isize)).flags = 0 as i32 as u16
+        (*disk.desc.offset(idx[1usize] as isize)).flags = 0 as i32 as u16
     } else {
         // device reads b->data
-        (*disk.desc.offset(idx[1 as i32 as usize] as isize)).flags = VRING_DESC_F_WRITE as u16
+        (*disk.desc.offset(idx[1usize] as isize)).flags = VRING_DESC_F_WRITE as u16
     }
 
-    let fresh0 = &mut (*disk.desc.offset(idx[1 as i32 as usize] as isize)).flags;
+    let fresh0 = &mut (*disk.desc.offset(idx[1usize] as isize)).flags;
     *fresh0 = (*fresh0 as i32 | VRING_DESC_F_NEXT) as u16;
-    (*disk.desc.offset(idx[1 as i32 as usize] as isize)).next = idx[2 as i32 as usize] as u16;
+    (*disk.desc.offset(idx[1usize] as isize)).next = idx[2usize] as u16;
 
-    disk.info[idx[0 as i32 as usize] as usize].status = 0 as i32 as libc::c_char;
-    (*disk.desc.offset(idx[2 as i32 as usize] as isize)).addr = &mut (*disk
+    disk.info[idx[0usize] as usize].status = 0 as i32 as libc::c_char;
+    (*disk.desc.offset(idx[2usize] as isize)).addr = &mut (*disk
         .info
         .as_mut_ptr()
         .offset(*idx.as_mut_ptr().offset(0 as i32 as isize) as isize))
     .status as *mut libc::c_char
         as usize;
-    (*disk.desc.offset(idx[2 as i32 as usize] as isize)).len = 1 as u32;
+    (*disk.desc.offset(idx[2usize] as isize)).len = 1 as u32;
     // device writes the status
-    (*disk.desc.offset(idx[2 as i32 as usize] as isize)).flags = VRING_DESC_F_WRITE as u16;
-    (*disk.desc.offset(idx[2 as i32 as usize] as isize)).next = 0 as i32 as u16;
+    (*disk.desc.offset(idx[2usize] as isize)).flags = VRING_DESC_F_WRITE as u16;
+    (*disk.desc.offset(idx[2usize] as isize)).next = 0 as i32 as u16;
 
     // record struct Buf for virtio_disk_intr().
     (*b).disk = 1;
-    disk.info[idx[0 as i32 as usize] as usize].b = b;
+    disk.info[idx[0usize] as usize].b = b;
 
     // avail[0] is flags
     // avail[1] tells the device how far to look in avail[2...].
@@ -301,7 +301,7 @@ pub unsafe fn virtio_disk_rw(mut b: *mut Buf, mut write: i32) {
     *disk
         .avail
         .offset((2 as i32 + *disk.avail.offset(1 as i32 as isize) as i32 % NUM) as isize) =
-        idx[0 as i32 as usize] as u16;
+        idx[0usize] as u16;
     ::core::intrinsics::atomic_fence();
     *disk.avail.offset(1 as i32 as isize) =
         (*disk.avail.offset(1 as i32 as isize) as i32 + 1 as i32) as u16;
@@ -313,8 +313,8 @@ pub unsafe fn virtio_disk_rw(mut b: *mut Buf, mut write: i32) {
     while (*b).disk == 1 {
         sleep(b as *mut libc::c_void, &mut disk.vdisk_lock);
     }
-    disk.info[idx[0 as i32 as usize] as usize].b = ptr::null_mut();
-    free_chain(idx[0 as i32 as usize]);
+    disk.info[idx[0usize] as usize].b = ptr::null_mut();
+    free_chain(idx[0usize]);
     disk.vdisk_lock.release();
 }
 pub unsafe fn virtio_disk_intr() {
